pyproject.toml
[tool.poetry]
name = "stockpredictor-streamlit"
version = "0.1.0"
description = ""
authors = ["MarkusSagen <markus.john.sagen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.1.0"
pystan = "2.19.1.1"
fbprophet = "^0.7.1"
yfinance = "^0.1.64"
download-yfinance-data = "^0.0.1"
plotly = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
